refator(alertDialog): refator confirm delete dialog to be reusable component
